+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - #60 - Added new provider `Seaborn` using `Html` and a new parser for Email Subject: `EmailSubjectParser`
 - #66 - Added new provider `Momentum` using `Html` and `EmailSubjectParser`
 - #61 - Added new provider `Colt` using `ICal` and `Csv`
+- #68 - Added new provider `HGC` using `Html` and `EmailSubjectParser`
 
 ### Fixed